pyproject.toml
[tool.poetry]
name = "cinereserve-api"
version = "0.1.0"
description = "High-performance RESTful backend for managing cinema operations"
authors = ["Samuel Gomes"]
readme = "README.md"
packages = [{include = "cinereserve"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
redis = "^5.0"
djangorestframework-simplejwt = "^5.3"
django-environ = "^0.11"
gunicorn = "^22.0"
django-cors-headers = "^4.3"
django-redis = "^5.4.0"
drf-spectacular = "^0.27.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
pytest-django = "^4.8"
pytest-cov = "^5.0"
factory-boy = "^3.3"
faker = "^24.0"
black = "^24.3"
isort = "^5.13"
flake8 = "^7.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cinereserve.settings"
python_files = ["tests.py", "test_*.py", "*_test.py"]
addopts = "--cov=. --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
